// src/lib/api/generated/model/subscriptionStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SubscriptionStatus = {
  trial: 'trial',
  active: 'active',
  past_due: 'past_due',
  grace: 'grace',
  blocked: 'blocked',
  cancelled: 'cancelled',
} as const;